View/Edit mode. Sharing passwords between users.

+++ PasswordWallet/Controllers/PasswordsController.cs
@@ -18,6 +18,12 @@ using Newtonsoft.Json;
 
 namespace PasswordWallet.Controllers
 {
+    public enum ViewMode
+    {
+        View,
+        Edit
+    }
+
     public class PasswordsController : Controller
     {
         private readonly IDbContext dbContext;
@@ -42,10 +48,57 @@ namespace PasswordWallet.Controllers
             }
         }
 
+        protected ViewMode GetViewMode()
+        {
+            ViewMode mode;
+
+            try
+            {
+                mode = (ViewMode)Enum.Parse(typeof(ViewMode), HttpContext.Session.GetString("ViewMode"));
+            }
+            catch
+            {
+                mode = ViewMode.View;
+            }
+
+            return mode;
+        }
+
+        protected void SetViewMode(ViewMode mode)
+        {
+            HttpContext.Session.SetString("ViewMode", mode.ToString());
+        }
+
+        [HttpPost]
+        public ActionResult ChangeMode()
+        {
+            var viewMode = GetViewMode();
+            
+            SetViewMode(
+                viewMode == ViewMode.View ? 
+                ViewMode.Edit : 
+                ViewMode.View);
+
+            return RedirectToAction(nameof(Index));
+        }
+
         // GET: Passwords
         public ActionResult Index()
         {
-            userInfo = JsonConvert.DeserializeObject<UserInfo>(HttpContext.Session.GetString("UserInfo"));
+            userInfo = GetUserInfo();
+
+            ViewBag.Message = GetViewMode();
+
+            try
+            {
+                string message = HttpContext.Session.GetString("WarningMessage");
+
+                if (message != "")
+                {
+                    ModelState.AddModelError("Warning", message);
+                }
+            }
+            catch { }
 
             return View(dbContext.GetPasswordListByUserId(userInfo.Id));
         }
@@ -53,9 +106,50 @@ namespace PasswordWallet.Controllers
         // GET: Passwords/ShowLog
         public ActionResult ShowLog()
         {
-            userInfo = JsonConvert.DeserializeObject<UserInfo>(HttpContext.Session.GetString("UserInfo"));
+            userInfo = GetUserInfo();
 
             return View(dbContext.GetLoginLogListByUserId(userInfo.Id));
+        }
+
+        protected UserInfo GetUserInfo()
+        {
+            UserInfo info;
+
+            try
+            {
+                info = JsonConvert.DeserializeObject<UserInfo>(HttpContext.Session.GetString("UserInfo"));
+            }
+            catch
+            {
+                info = new UserInfo();
+            }
+            return info;
+        }
+
+        // GET: Passwords/SharePassword
+        public ActionResult SharePassword(string passwordHash)
+        {
+            if (IsSharedPassword(
+                passwordHash, 
+                "You cannot share this password! You are not an owner."))
+            {
+                return RedirectToAction(nameof(Index));
+            }
+            else
+            {
+                var password = dbContext.GetPasswordByHash(passwordHash);
+
+                HttpContext.Session.SetString(
+                    "PasswordShareInfo",
+                    JsonConvert.SerializeObject(
+                        new PasswordShareInfo()
+                        {
+                            OwnerId = GetUserInfo().Id,
+                            PasswordId = password.Id,
+                        }));
+
+                return RedirectToAction(nameof(Index), "PasswordShare");
+            }
         }
 
         // GET: Passwords/Create
@@ -70,7 +164,7 @@ namespace PasswordWallet.Controllers
         {
             try
             {
-                userInfo = JsonConvert.DeserializeObject<UserInfo>(HttpContext.Session.GetString("UserInfo"));
+                userInfo = GetUserInfo();
 
                 UserModel user;
 
@@ -102,19 +196,105 @@ namespace PasswordWallet.Controllers
         // GET: Passwords/Details/5
         public ActionResult Details(string passwordHash)
         {
-            userInfo = JsonConvert.DeserializeObject<UserInfo>(HttpContext.Session.GetString("UserInfo"));
+            userInfo = GetUserInfo();
 
             PasswordModel password = dbContext.GetPasswordByHash(passwordHash);
-            
-            password.PasswordHash = EncryptionHelper.DecryptPasswordAES(password.PasswordHash, userInfo.LoggedUserPassword);
+            if (password != null)
+            {
+                password.PasswordHash = EncryptionHelper.DecryptPasswordAES(password.PasswordHash, userInfo.LoggedUserPassword);
 
-            return View(password);
+                return View(password);
+            }
+            else
+            {
+                return View(
+                    "EnterSharingKey",
+                    new SharePasswordKey()
+                    {
+                        KeyHash = passwordHash
+                    });
+            }
+        }
+
+        public bool IsSharedPassword(string passwordHash, string message = "")
+        {
+            if (dbContext.GetSharedPasswordByHash(passwordHash) == null)
+            {
+                HttpContext.Session.SetString(
+                    "WarningMessage",
+                    ""
+                    );
+
+                return false;
+            }
+            else
+            {
+                HttpContext.Session.SetString(
+                    "WarningMessage",
+                    message
+                    );
+
+                return true;
+            }
+        }
+
+        [HttpPost]
+        public ActionResult SharePasswordDatail(SharePasswordKey sharePasswordKey)
+        {
+            var sharedPassword = dbContext.GetSharedPasswordByHash(sharePasswordKey.KeyHash);
+
+            try
+            {
+                sharedPassword.PasswordHash =
+                    EncryptionHelper.DecryptPasswordAES(
+                        sharedPassword.PasswordHash,
+                        sharePasswordKey.Key
+                        );
+
+                if (ValidateDecryptedPassword(sharedPassword.PasswordHash))
+                {
+                    throw new Exception();
+                }
+
+                HttpContext.Session.SetString(
+                    "WarningMessage",
+                    ""
+                    );
+
+                return View("SharedPasswordDetails", sharedPassword);
+            }
+            catch 
+            {
+                HttpContext.Session.SetString(
+                    "WarningMessage",
+                    "Wrong sharing key!"
+                    );
+
+                return RedirectToAction(nameof(Index));
+            }
+        }
+
+        public bool ValidateDecryptedPassword(string input)
+        {
+            const int MaxAnsiCode = 255;
+
+            return input.Any(c => c > MaxAnsiCode);
         }
 
         // GET: Passwords/Delete/5
         public ActionResult Delete(string passwordHash)
         {
-            return View(dbContext.GetPasswordByHash(passwordHash));
+            if (ValidateViewMode("Cannot delete. You are in view mode!") &&
+                !IsSharedPassword(
+                    passwordHash,
+                    "You cannot delete this password! You are not an owner."))
+            {
+                return View(dbContext.GetPasswordByHash(passwordHash));
+            }
+            else
+            {
+                return RedirectToAction(nameof(Index));
+            }
         }
 
         // POST: Passwords/Delete/5
@@ -124,9 +304,12 @@ namespace PasswordWallet.Controllers
         {
             try
             {
-                string passwordHash = collection["PasswordHash"];
+                if (ValidateViewMode("Cannot delete. You are in view mode!"))
+                {
+                    string passwordHash = collection["PasswordHash"];
 
-                dbContext.DeletePassword(passwordHash);
+                    dbContext.DeletePassword(passwordHash);
+                }
 
                 return RedirectToAction(nameof(Index));
             }
@@ -134,6 +317,32 @@ namespace PasswordWallet.Controllers
             {
                 return View();
             }
+        }
+
+        protected bool ValidateViewMode(string message = "")
+        {
+            bool isOk;
+
+            if (GetViewMode() == ViewMode.Edit)
+            {
+                HttpContext.Session.SetString(
+                    "WarningMessage",
+                    ""
+                    );
+
+                isOk = true;
+            }
+            else
+            {
+                HttpContext.Session.SetString(
+                    "WarningMessage",
+                    message
+                    );
+
+                isOk = false;
+            }
+
+            return isOk;
         }
 
     }
